Replace `supertask` with `base` for moved modules
